requirements.txt
# Everything needed in production

csv
BeautifulSoup4
werkzeug<1.0 # This can be removed when switching from Flask-OAuthlib
Flask>=0.10
Flask-DebugToolbar
Flask-Bootstrap
path.py
Flask-OAuthlib # This is deprecated and can not use werkzeug>=1.0
Flask-Markdown
Flask-WTF
Flask-Mail
Flask-WeasyPrint
coloredlogs
query_phenomizer
Flask-Babel
livereload
python-dateutil
pymongo<3.7
pathlib
WeasyPrint==0.42.3
xlsxwriter
click<7.0
requests-oauthlib<1.2.0
oauthlib<3.0.0
flask_ldap3_login

# unclear
Flask-Login

# Parsing
cyvcf2<0.10.0
PyYaml>=5.1
ped_parser

# extras
intervaltree==3.0.2

# chanjo-report
python-dateutil

# Requesting resources

# Required by scripts
tabulate
